Centralize pipeline building logics in Buildkite::Builder::Pipeline
